+++ app/src/features/today-quest/components/template/TodayQuests/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import {AppText, FlexGapContainer} from '@shared/components/base';
 import useUserStore from '@shared/stores/useUserStore';
-import useTodayQuestQuery from 'features/quest/hooks/useTodayQuestQuery';
-import usePopupRenewTodayDialog from 'features/quest/hooks/usePopupRenewTodayDialog';
+import useTodayQuestQuery from 'features/today-quest/hooks/useTodayQuestQuery';
+import usePopupRenewTodayDialog from 'features/today-quest/hooks/usePopupRenewTodayDialog';
 import {TodayQuestItem} from '../../module';
 
 function TodayQuests() {
