+++ app/src/features/today-quest/components/template/index.ts
